v1.4.0 (see desc. for release plan)
i'm going to be calling v1.4.0 a "dev alpha" build because it's still in development and in its alpha phase, not stable at all yet. this is going to be a big release, so thats why I'm skipping 1.3.0. if this release takes too much time, i'll publish 1.3.0 with some of the 1.4.0 planned features and release 1.4.0 with the rest, but for now I'm not really going to worry about that and just develop it because it will take some time. do not use the dev version daily as it's going to be really crazy around here, but please do try it and provide feedback. thanks!

-jackson